+++ resources/js/components/System/Groups.vue
@@ -53,30 +53,30 @@ export default {
         SystemGroupEditor,
     },
     mounted () {
-        this.$store.dispatch('loadGroups');
-        this.$store.dispatch('loadUsers');
+        this.$store.dispatch('systemGroups/load');
+        this.$store.dispatch('systemUsers/load');
     },
     computed: {
         ...mapState({
-            editing: state => state.Group.editing,
-            search: state => state.Group.currentFilter,
-            showSysGroups: state => state.Group.showSystem,
+            editing: state => state.systemGroups.editing,
+            search: state => state.systemGroups.currentFilter,
+            showSysGroups: state => state.systemGroups.showSystem,
         }),
-        ...mapGetters([
-            'groups',
-            'filteredGroups',
-        ]),
+        ...mapGetters({
+            groups: 'systemGroups/all',
+            filteredGroups: 'systemGroups/filtered',
+        }),
         listWidth() {
             return this.editing ? 10 : 16;
         },
     },
     methods: {
         ...mapMutations({
-            filterGroups: 'setFilter',
-            toggleSysGroups: 'toggleSystemGroups',
+            filterGroups: 'systemGroups/setFilter',
+            toggleSysGroups: 'systemGroups/toggleSystemGroups',
         }),
         ...mapActions({
-            edit: 'editGroup',
+            edit: 'systemGroups/edit',
         }),
     },
 }