+++ resources/js/pages/Files/Browser.vue
@@ -39,11 +39,11 @@ import FileList from '../../components/Files/Browser/FileList';
 
 export default {
     mounted () {
-        this.$store.dispatch('loadSites');
-        this.$store.dispatch('loadFiles', { path: this.path });
+        this.$store.dispatch('sites/load');
+        this.$store.dispatch('files/load', { path: this.path });
     },
     beforeRouteUpdate (to, from, next) {
-        this.$store.dispatch('loadFiles', { path: to.params.path });
+        this.$store.dispatch('files/load', { path: to.params.path });
         next();
     },
     props: [
@@ -53,11 +53,11 @@ export default {
         FileList,
     },
     computed: {
-        ...mapGetters([
-            'currentPath',
-            'getSiteByDocroot',
-            'files',
-        ]),
+        ...mapGetters({
+            currentPath: 'files/currentPath',
+            findSite: 'sites/findByDocroot',
+            files: 'files/all',
+        }),
         pathParts: function() {
             let parts = [],
                 path = '';
@@ -74,7 +74,7 @@ export default {
             return parts;
         },
         site: function() {
-            return this.getSiteByDocroot(this.currentPath);
+            return this.findSite(this.currentPath);
         },
     },
     methods: {
@@ -94,7 +94,7 @@ export default {
             next = next ? next : '/';
 
             this.$router.push({ name: 'files', params: { path: next } });
-            this.$store.dispatch('loadFiles', { path: next });
+            this.$store.dispatch('files/load', { path: next });
         },
     }
 }